+++ accounts/models.py
@@ -2,6 +2,7 @@ from django.db import models
 from django.contrib.auth.models import AbstractUser
 from imagekit.models import ProcessedImageField
 from imagekit.processors import ResizeToFill
+from movies.models import Movie
 
 # Create your models here.
 class User(AbstractUser):
@@ -10,4 +11,5 @@ class User(AbstractUser):
                             format='JPEG',
                             options={'quality': 100}
                 )
+    recommended_movie = models.ManyToManyField(Movie)
