+++ accounts/templates/accounts/profile.html
@@ -66,8 +66,10 @@
   
   <div class="card w-100">
     <div class="card-body">
-      <p class="card-title">평점 분포</p>
-      <h4 class="card-text text-center">(공 란)</h4>
+      <p class="card-title">추천받은 영화</p>
+      {% for recommended_movie in recommended_movies %}
+        <h4 class="card-text text-center">{{ recommended_movie.title }}</h4>
+      {% endfor %}
     </div>
   </div>
