--- movies/templates/movies/worldcup2.html
@@ -1,136 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-  <h2>WorldCup</h2>
-  <h4 id="round" data-last-page="{{ last_page }}">{{ init_page }} / {{ last_page }}</h4>
-  <h4>내가 선택한 영화: </h4>
-  <hr>
-  <div class="row">
-    {% for movie in movies %}
-      <div class="movie-images col-6">
-        <img class="img" id="img-{{ movie.pk }}" data-movie-pk="{{ movie.pk }}" src="https://image.tmdb.org/t/p/original/{{ movie.poster_path }}"alt="">
-      </div>
-    {% endfor %}
-  </div>
-{% endblock content %}
-
-{% block script %}
-  <script>
-  let pageNum = 1
-  let lastPage = 0
-  let totalRound = 0
-  let seletedMoviePks = []
-
-  window.onload = function() {
-    lastPage = Number(document.querySelector('#round').dataset.lastPage)
-    totalRound = 2 * lastPage
-  }
-
-  const parentImages = document.querySelectorAll('.movie-images')
-  parentImages.forEach((image) => {
-    image.addEventListener('click', (event) => {
-      if (lastPage === 1) {
-        seletedMoviePks.push(event.target.dataset.moviePk)
-        console.log('finished')
-        window.location.href=`/movies/worldcup/end/?seletedMoviePks=${seletedMoviePks[seletedMoviePks.length - 1]}&totalRound=${totalRound}` 
-        //  axios({
-        //   method: 'get',
-        //   url: "/movies/worldcup/end/",
-        //   params: {
-        //     seletedMoviePks: seletedMoviePks[seletedMoviePks.length - 1],
-        //     totalRound: totalRound,
-        //   }
-        // })
-      } 
-      else if(pageNum === lastPage) {
-        // next_stage request
-        seletedMoviePks.push(event.target.dataset.moviePk)
-        console.log('last')
-        axios({
-          method: 'get',
-          url: `/movies/worldcup/next/`,
-          params: {
-            page: `${pageNum+1}`,
-            seletedMoviePks: seletedMoviePks[seletedMoviePks.length - 1],
-          }
-        }) .then((res) => {
-          console.log('응답 성공')
-          lastPage = res.data.last_page
-
-
-          movies = JSON.parse(res.data.movies)
-          // console.log(movies)
-          pageNum = 1
-          const round = document.querySelector('#round')
-          round.innerText = `${pageNum} / ${lastPage}`
-          
-
-          const images = document.querySelectorAll('.img')
-
-          images.forEach(function(image, index) {
-            preMoviePk = image.getAttribute('data-movie-pk')
-
-            moviePk = movies[index].pk
-            moviePosterPath = movies[index].fields.poster_path
-            const image1 = document.querySelector(`#img-${preMoviePk}`)
-            const newImg = document.createElement('img')
-            newImg.classList.add('img')
-            newImg.src = `https://image.tmdb.org/t/p/original/${moviePosterPath}`
-            newImg.id = `img-${moviePk}`
-            newImg.dataset.moviePk = `${moviePk}`
-
-            const parent = image1.parentNode
-            console.log(newImg)
-            parent.replaceChild(newImg, image1)
-          })
-        })
-      } else {
-
-      ///////////////////////////////////////////////////////
-      // next_page request
-      seletedMoviePks.push(event.target.dataset.moviePk)
-      axios({
-        method: 'get',
-        url: `/movies/worldcup/next/`,
-        headers: {'x-requested-with': 'XMLHttpRequest'},
-        params: {
-          page: `${pageNum+1}`,
-          seletedMoviePks: seletedMoviePks[seletedMoviePks.length - 1],
-        }
-      })
-        .then((res) => {
-          console.log(res)
-          const movies = res.data
-          moviePk = movies[0].pk
-          moviePosterPath = movies[0].fields.poster_path
-
-          const images = document.querySelectorAll('.img')
-
-          images.forEach(function(image, index) {
-            preMoviePk = image.getAttribute('data-movie-pk')
-
-            moviePk = movies[index].pk
-            moviePosterPath = movies[index].fields.poster_path
-            const image1 = document.querySelector(`#img-${preMoviePk}`)
-            const newImg = document.createElement('img')
-            newImg.classList.add('img')
-            newImg.src = `https://image.tmdb.org/t/p/original/${moviePosterPath}`
-            newImg.id = `img-${moviePk}`
-            newImg.dataset.moviePk = `${moviePk}`
-
-            const parent = image1.parentNode
-            parent.replaceChild(newImg, image1)
-          })
-          const round = document.querySelector('#round')
-          round.innerText = `${pageNum+1} / ${lastPage}`
-          pageNum += 1
-        })
-      }
-
-    })
-  })
-
-  </script>
-{% endblock script %}
-
-
