+++ movies/templates/movies/worldcup_end.html
@@ -1,6 +1,8 @@
 {% extends 'base.html' %}
 {% block content %}
-  <h1> 영화 월드컵 {{ total_round }}강 우승 영화</h1>
+  <h2 class="text-center"> 영화 월드컵 {{ total_round }}강 우승 영화</h2>
+  <h2 class="text-center">{{ movie.title }}</h2>
+  
   <div class="row">
     <div class="col-3">
     </div>
