+++ movies/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from . import views, views2
+from . import views
 
 app_name = 'movies'
 
@@ -12,8 +12,9 @@ urlpatterns = [
     path('<int:movie_pk>/rank/', views.create_rank, name='create_rank'),
     path('<int:movie_pk>/rank/<int:rank_pk>/update', views.update_rank, name='update_rank'),
     path('<int:movie_pk>/rank/<int:rank_pk>/delete', views.delete_rank, name='delete_rank'),
-    path('worldcup/', views2.worldcup_init, name="worldcup_init"),
-    path('worldcup/next/', views2.worldcup_next_round, name="worldcup_next_round"),
-    path('worldcup/end/', views2.worldcup_end, name="worldcup_end"),
+    path('worldcup/', views.worldcup_init, name="worldcup_init"),
+    path('worldcup/next/', views.worldcup_next_round, name="worldcup_next_round"),
+    path('worldcup/end/', views.worldcup_end, name="worldcup_end"),
+    path('worldcup/select_round/',views.worldcup_select_round, name="worldcup_select_round"),
 ]
