+++ templates/_nav.html
@@ -10,6 +10,7 @@
       <ul class="navbar-nav me-auto mb-2 mb-lg-0">
         <a class="nav-link" href="{% url 'movies:index' %}">영화목록</a>
         <a class="nav-link" href="{% url 'community:index' %}">커뮤니티</a>
+        <a class="nav-link" href="{% url 'movies:worldcup_select_round' %}">영화월드컵</a>
       </ul>
 
       <form class="d-flex">
